made use of ContextPropagatingThreadFactory and made service stop if idle
ContextPropagatingThreadFactory sounds like it might help fix the database corruption as I suspect that competing processes are writing to it, not aware of oneanother, which should be the case if they don't share the Context.
Stopping the service was removed but it makes sense in order to be able to start it again when nothing happens.