some bugfixes and refactorings
passphrase as List, not ArrayList
accountIndexStrings as ConcurrentSkipListSet to fix concurrency crash
string constants should never be stored in res/values/strings.xml. That file is exclusively for localizable strings.
changed Preferences.getStringSet to Preferences.getString for the passphrase, as a Set<String> would firstly not necessarily have an ordering and secondly not allow duplicates which the passphrase clearly allows. storing now as "word1 word2 word3 ..."